requirements.txt
pyserial~=3.5
numpy~=2.0.0rc1
PyQt6~=6.6.1
pyqtgraph~=0.13.4
matplotlib~=3.9.0rc2